move INHERITABLE_PROPERTIES to roosterjs-editor-types
