+++ packages/roosterjs-editor-types/lib/Constants.ts
@@ -1,0 +1,9 @@
+// Inheritable CSS properties
+// Ref: https://www.w3.org/TR/CSS21/propidx.html
+export const INHERITABLE_PROPERTIES = (
+    'border-spacing,caption-side,color,' +
+    'cursor,direction,empty-cells,font-family,font-size,font-style,font-variant,font-weight,' +
+    'font,letter-spacing,line-height,list-style-image,list-style-position,list-style-type,' +
+    'list-style,orphans,quotes,text-align,text-indent,text-transform,visibility,white-space,' +
+    'widows,word-spacing'
+).split(',');
